# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Signalement Lalantsika",
        "description": "Documentation de l'API pour la gestion des signalements et entreprises.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API principale"
        }
    ],
    "paths": {
...
        "/reports": {
            "get": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Liste tous les signalements",
                "description": "Liste tous les signalements avec leurs relations",
                "operationId": "0bc0c4872b152b2b9b2ad92ea69f5a88",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reports/{id}": {
            "get": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Détails d'un signalement",
                "description": "Récupère un signalement par son ID",
                "operationId": "d5ed27e574097430c99794f504d8d151",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails récupérés avec succès"
                    },
                    "404": {
                        "description": "Signalement non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Mettre à jour un signalement (budget, entreprise)",
                "description": "Met à jour un signalement (budget, entreprise, etc.)",
                "operationId": "00adfff7b6301d610073660350d33538",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_entreprise": {
                                        "type": "integer"
                                    },
                                    "budget": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mise à jour réussie"
                    },
                    "404": {
                        "description": "Signalement non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Supprimer un signalement",
                "description": "Supprime un signalement",
                "operationId": "de9f82153c925777d11ecb19839a836f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Signalement supprimé"
                    },
                    "404": {
                        "description": "Signalement non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reports/{id}/histostatut": {
            "get": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Récupérer l'historique des statuts",
                "description": "Récupère l'historique des statuts d'un signalement",
                "operationId": "6ecfb4e06fb56e952916e93c44e79dda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historique récupéré"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Ajouter une étape d'historique (changement de statut)",
                "description": "Ajoute un historique de statut (histostatut) pour un signalement",
                "operationId": "c0b5895cee1a4ff5707930bb4d202df7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id_statut",
                                    "description",
                                    "daty"
                                ],
                                "properties": {
                                    "id_statut": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "daty": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Historique ajouté avec succès"
                    },
                    "404": {
                        "description": "Signalement non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/statuses": {
            "get": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Liste des statuts de signalement possibles",
                "description": "Liste tous les statuts disponibles",
                "operationId": "85119adaf6ac9a1f5990f9645fda494f",
                "responses": {
                    "200": {
                        "description": "Liste des statuts"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}